pyproject.toml
[tool.black]
exclude = '''
/(
  | __pycache__
  | alembic
)/
'''

[tool.flake8]
extend-ignore = 'F403, F405'
max-line-length = 88
exclude = '''
/(
  | __pycache__
  | alembic
)/
'''

[tool.pylint]
# Use multiple processes to speed up Pylint. Specifying 0 will auto-detect the
# number of processors available to use.
jobs=0

[tool.mypy]
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
no_implicit_optional = true
ignore_missing_imports = true
no_site_packages = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
strict_equality = true

[tool.poetry]
name = "pymultinotifier"
version = "0.1.0"
description = ""
authors = ["Tutor Exilius <tutorexilius@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
flet = "^0.1.64"
SQLAlchemy = "^1.4.43"
alembic = "^1.8.1"
pyserde = "^0.9.3"
toml = "^0.10.2"
tomli-w = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"
mypy = "^0.990"
pylint = "^2.15.5"
flake9 = "^3.8.3"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
